% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{calculate_edit_distance}
\alias{calculate_edit_distance}
\title{Vectorized Levenshtein (edit) distance}
\usage{
calculate_edit_distance(str1_vec, str2_vec, ignore_strings = NULL)
}
\arguments{
\item{str1_vec}{A character vector.}

\item{str2_vec}{A character vector, must be the same length as `str1_vec`.}

\item{ignore_strings}{A character vector of phrases to remove from
both strings (case-insensitive) before comparison.}
}
\value{
A numeric vector of edit distances.
}
\description{
Calculates the edit distance for two vectors of strings,
element by element.
}
