% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{calculate_tfidf_similarity}
\alias{calculate_tfidf_similarity}
\title{Vectorized TF-IDF weighted similarity score}
\usage{
calculate_tfidf_similarity(str1_vec, str2_vec, tfidf_weights)
}
\arguments{
\item{str1_vec}{A character vector.}

\item{str2_vec}{A character vector, same length as `str1_vec`.}

\item{tfidf_weights}{A data.frame with columns `word` and `idf`.}
}
\value{
A numeric vector of similarity scores.
}
\description{
Calculates a similarity score for two vectors of strings based on
the sum of IDF weights for their common words.
}
