% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{clean_but_keep}
\alias{clean_but_keep}
\title{Remove all objects from an environment *except* specified objects and functions}
\usage{
clean_but_keep(..., env = .GlobalEnv)
}
\arguments{
\item{...}{Unquoted names of objects (dataframes, vectors, etc.) to keep.}

\item{env}{The environment to clean. Defaults to the Global Environment.}
}
\value{
NULL. This function modifies the environment as a side effect.
}
\description{
This function is a "safe" version of `rm(list = ls())`. It finds all
objects in the specified environment, removes the objects you list to "keep"
and all functions, and then deletes what is left.
}
