% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{count_consecutive_overlap}
\alias{count_consecutive_overlap}
\title{Vectorized Longest Common Substring (LCS)}
\usage{
count_consecutive_overlap(str1_vec, str2_vec, ignore_strings = NULL)
}
\arguments{
\item{str1_vec}{A character vector.}

\item{str2_vec}{A character vector, must be the same length as `str1_vec`.}

\item{ignore_strings}{A character vector of phrases to remove from
both strings (case-insensitive) before comparison.}
}
\value{
A numeric vector of LCS lengths.
}
\description{
Calculates the length of the longest common substring for two
vectors of strings, element by element.
}
