% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{find_best_match}
\alias{find_best_match}
\title{Find the best (longest) substring match from an approved list}
\usage{
find_best_match(tc_name, approved_list)
}
\arguments{
\item{tc_name}{A single string to be checked.}

\item{approved_list}{A character vector of "approved" names to match against.}
}
\value{
The longest matching string from `approved_list`, or the
  original `tc_name` if no match is found.
}
\description{
Compares a single name (`tc_name`) against a list of approved names.
It returns a perfect match if one exists. If not, it finds all
approved names that are substrings of `tc_name` and returns the
longest one.
}
