% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{lazy_hist}
\alias{lazy_hist}
\title{Creates a publication-quality histogram with sensible defaults.}
\usage{
lazy_hist(df, var_string, binwidth_val = 30)
}
\arguments{
\item{df}{The dataframe containing the data.}

\item{var_string}{The *string name* of the numeric variable to plot.}

\item{binwidth_val}{The width of the histogram bins.}
}
\value{
A ggplot object representing the histogram.
}
\description{
Creates a publication-quality histogram with sensible defaults.
}
