% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{lunique}
\alias{lunique}
\title{Get the count of unique values}
\usage{
lunique(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A single numeric value representing the count of unique elements.
}
\description{
Calculates the length of the vector returned by `unique()`.
}
