% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{mudate}
\alias{mudate}
\title{Mutate multiple columns to Date objects}
\usage{
mudate(df, ...)
}
\arguments{
\item{df}{A data.frame.}

\item{...}{One or more unquoted column names to be converted to Date.}
}
\value{
A data.frame with the specified columns mutated to Date.
}
\description{
A wrapper for `mutate(across(c(...), as.Date))` to quickly convert
one or more columns to the Date class.
}
