% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{na_if_true}
\alias{na_if_true}
\title{Conditionally replace values with NA for multiple columns.}
\usage{
na_if_true(.data, ...)
}
\arguments{
\item{.data}{The dataframe to modify.}

\item{...}{A series of named arguments in the format `column_name = condition`.}
}
\value{
A dataframe with values replaced by NA where conditions are met.
}
\description{
Conditionally replace values with NA for multiple columns.
}
