% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{na_replace}
\alias{na_replace}
\title{Replace values with NA based on a condition for a single column}
\usage{
na_replace(.data, var, condition)
}
\arguments{
\item{.data}{A data.frame.}

\item{var}{The unquoted name of the column to modify.}

\item{condition}{A logical expression. Where this is `TRUE`,
the value in `var` will be set to `NA`.}
}
\value{
A data.frame with the specified column modified.
}
\description{
Dynamically selects a column and replaces its values with `NA`
(of the correct type) where a given condition is `TRUE`.
}
