% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{nable}
\alias{nable}
\title{Create a table of NA vs. non-NA values}
\usage{
nable(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A `table` object with counts for `TRUE` (is NA) and `FALSE` (is not NA).
}
\description{
Create a table of NA vs. non-NA values
}
