% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{nable_pct}
\alias{nable_pct}
\title{Print NA counts and percentages for a dataframe}
\usage{
nable_pct(x)
}
\arguments{
\item{x}{A data.frame or tibble.}
}
\value{
NULL. This function prints its results to the console.
}
\description{
Print NA counts and percentages for a dataframe
}
\details{
The function prints the `nable()` count and then divides this
  by the total number of rows in the dataframe.
}
