% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{rm_any_of}
\alias{rm_any_of}
\title{Safely remove a vector of objects from the environment}
\usage{
rm_any_of(objects_to_remove)
}
\arguments{
\item{objects_to_remove}{A character vector of object names to delete.}
}
\value{
NULL. This function modifies the environment as a side effect.
}
\description{
Takes a character vector of object names and removes only those that
currently exist in the Global Environment, without throwing an error
for objects that do not exist.
}
