% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCY_pack_rn.R
\name{table2}
\alias{table2}
\title{Create a "table of tables"}
\usage{
table2(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A `table` object showing the frequency of frequencies.
}
\description{
Reports the frequencies of frequencies. For example, if a vector has
three 1s, two 5s, and two 6s, this function will report that
two values (5 and 6) appear 2 times, and one value (1) appears 3 times.
}
